+++ src/ACESVGjsPanel.tsx
@@ -10,21 +10,21 @@ interface MappedElements {
 interface Props extends PanelProps<ACESVGOptions> { }
 
 interface PanelState {
-  addAllIDs: boolean;
-  svgNode: SVGElement | SVGDom | null;
-  svgSource: string | null;
-  mappedElements: MappedElements | null;
-  svgMappings: SVGIDMapping[];
-  initFunctionSource: string;
-  initFunction: Function | null;
-  eventFunctionSource: string;
-  eventFunction: Function | null;
-  initialized: boolean;
-  context: any;
+  readonly addAllIDs: boolean;
+  readonly svgNode: SVGElement | SVGDom | null;
+  readonly svgSource: string | null;
+  readonly mappedElements: MappedElements | null;
+  readonly svgMappings: SVGIDMapping[];
+  readonly initFunctionSource: string;
+  readonly initFunction: Function | null;
+  readonly eventFunctionSource: string;
+  readonly eventFunction: Function | null;
+  readonly initialized: boolean;
+  readonly context: any;
 }
 
 interface TextMappedElement extends SVGElement {
-  textElement: Element;
+  readonly textElement: Element;
 }
 
 SVGExtend(SVGElement, {
@@ -34,7 +34,6 @@ SVGExtend(SVGElement, {
   animateContRotate: function (this: SVGElement, speed: number) {
     return (
       this.animate(speed)
-        //@ts-ignore
         .ease('-')
         //@ts-ignore
         .rotate(360)
@@ -93,7 +92,6 @@ SVGExtend(SVGDom, {
   },
 });
 
-// export class SimplePanel extends PureComponent<Props, State> = ({ options, data, width, height }) => {
 export class ACESVGPanel extends PureComponent<Props, PanelState> {
   constructor(props: Props) {
     super(props);
@@ -112,7 +110,7 @@ export class ACESVGPanel extends PureComponent<Props, PanelState> {
     };
   }
 
-  initializeMappings(svgNode: SVGElement | SVGDom) {
+  private initializeMappings(svgNode: SVGElement | SVGDom): void {
     const svgMappings = this.props.options.svgMappings,
       currentElements: MappedElements = {};
     for (let i = 0; i < svgMappings.length; i++) {
@@ -125,10 +123,10 @@ export class ACESVGPanel extends PureComponent<Props, PanelState> {
     this.setState({ mappedElements: currentElements });
   }
 
-  mapAllIDs(svgNode: SVGDom) {
+  private mapAllIDs(svgNode: SVGDom): void {
     const svgMappings: SVGIDMapping[] = [...this.props.options.svgMappings];
     const nodeFilterID: NodeFilter = {
-      acceptNode: (node: Element) => {
+      acceptNode(node: Element) {
         if (node.id) {
           if (node.id !== '') {
             return NodeFilter.FILTER_ACCEPT;
@@ -153,7 +151,7 @@ export class ACESVGPanel extends PureComponent<Props, PanelState> {
     this.forceUpdate();
   }
 
-  mappingClickHandler(event: React.MouseEvent<HTMLElement, MouseEvent>) {
+  private mappingClickHandler(event: React.MouseEvent<HTMLElement, MouseEvent>): void {
     if (event.target) {
       let clicked = event.target as Element;
       let loopCount = 0;
@@ -180,7 +178,7 @@ export class ACESVGPanel extends PureComponent<Props, PanelState> {
     }
   }
 
-  renderSVG(element: SVGSVGElement | SVGDom | null) {
+  private renderSVG(element: SVGSVGElement | SVGDom | null): string | null {
     if (element) {
       if (
         this.props.options.initSource !== this.state.initFunctionSource ||
@@ -257,7 +255,7 @@ export class ACESVGPanel extends PureComponent<Props, PanelState> {
     }
   }
 
-  render() {
+  public render(): React.JSX.Element {
     return (
       <div
         onClick={this.props.options.captureMappings ? this.mappingClickHandler.bind(this) : undefined}